# Optimize the optimizer