--- a/topside/visualization/layout.py
+++ b/topside/visualization/layout.py
@@ -1,6 +1,3 @@
-from dataclasses import dataclass
-
-import numpy as np
 import networkx as nx