--- a/topside/visualization/layout.py
+++ b/topside/visualization/layout.py
@@ -2,6 +2,28 @@ import networkx as nx
 
 
 def terminal_graph(plumbing_engine):
+    """
+    Create an augmented graph suitable for visualization.
+
+    The terminal graph T is created by the following process, for a
+    plumbing engine P with graph G:
+        1. All non-atm nodes from G are placed in T.
+        2. For each component C in P:
+            a) For each node n in C:
+                i. Create a node C_n in T, where the node name of C_n is
+                   <name of C>.<name of n>.
+                ii. Create an edge in T from C_n to C.mapping[n].
+
+    Parameters
+    ----------
+
+    plumbing_engine: topside.PlumbingEngine
+
+    Returns
+    -------
+
+    t: networkx.Graph
+    """
     t = nx.Graph()
 
     for name, c in plumbing_engine.component_dict.items():