… topside/visualization/optimization/optimization.py
… topside/visualization/optimization/tests/test_cost_terms.py
… topside/visualization/optimization/tests/test_optimization_utils.py