Vectorize the graph layout optimization routine. Remove nested for loops and redundant calculations between cost terms and across iterations.

Speeds up `layout_plumbing_engine` by ~4.8x.

Current master:
```
mean time: 4.4294932126998905 seconds
max time: 4.631882905960083 seconds
min time: 4.306760311126709 seconds
```

This PR:
```
mean time: 0.9118092298507691 seconds
max time: 0.9235303401947021 seconds
min time: 0.8985953330993652 seconds
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/waterloo-rocketry/topside/17)
<!-- Reviewable:end -->