@wendi-yu 2020-05-19
Yep everything makes sense, I'll do the plumbing engine comments in the same style then. Preemptive LGTM if it's just comment `terminal_graph` from here :0